separate initializer
